refactor: extract shared cli support layer
