+++ src/cli.ts
@@ -12,7 +12,7 @@
  *   agentlog codex-notify             — legacy Codex notify handler
  */
 
-import { existsSync, readFileSync, rmSync } from "fs";
+import { existsSync, rmSync } from "fs";
 import { join, resolve } from "path";
 import { homedir } from "os";
 import { spawnSync } from "child_process";
@@ -21,7 +21,15 @@ import type { AgentLogConfig } from "./types.js";
 import { detectVaults, detectCli } from "./detect.js";
 import { isVersionAtLeast, MIN_CLI_VERSION, resolveCliBin, parseCliVersion } from "./obsidian-cli.js";
 import { registerHook, unregisterHook, isHookRegistered, CLAUDE_SETTINGS_PATH } from "./claude-settings.js";
-import { Errors, formatError } from "./errors.js";
+import {
+  ask,
+  detectBinary,
+  printObsidianCliStatus,
+  printSavedConfig,
+  saveMergedConfig,
+  validateVault,
+  validateVaultOrExit,
+} from "./cli-shared.js";
 import {
   CODEX_HOOKS_PATH,
   type CodexHookMutationResult,
@@ -35,92 +43,7 @@ import {
   unregisterCodexNotify,
 } from "./codex-settings.js";
 import { formatVersionHeadline, formatVersionOutput, getRuntimeInfo, readVersion, resolvePackageRoot } from "./version-info.js";
-import * as readline from "readline";
 import { Command } from "commander";
-
-function ask(prompt: string): Promise<string> {
-  const rl = readline.createInterface({ input: process.stdin, output: process.stdout });
-  return new Promise((resolve) => {
-    rl.question(prompt, (answer) => {
-      rl.close();
-      resolve(answer.trim());
-    });
-  });
-}
-
-function validateVault(vaultArg: string, plain: boolean): { vault: string; error?: string } {
-  const vault = resolve(expandHome(vaultArg));
-
-  if (!existsSync(vault)) {
-    return { vault, error: formatError(Errors.VAULT_NOT_FOUND(vault)) };
-  }
-
-  if (!plain) {
-    const obsidianDir = join(vault, ".obsidian");
-    if (!existsSync(obsidianDir)) {
-      return { vault, error: formatError(Errors.VAULT_NOT_OBSIDIAN(vault)) };
-    }
-  }
-
-  return { vault };
-}
-
-function validateVaultOrExit(vaultArg: string, plain: boolean): string {
-  const { vault, error } = validateVault(vaultArg, plain);
-  if (error) {
-    console.error(error);
-    process.exit(1);
-  }
-  return vault;
-}
-
-function saveMergedConfig(
-  vault: string,
-  plain: boolean,
-  extra: Partial<AgentLogConfig> = {}
-): AgentLogConfig {
-  const existing = loadConfig() ?? { vault };
-  const next: AgentLogConfig = {
-    ...existing,
-    ...extra,
-    vault,
-  };
-
-  if (plain) {
-    next.plain = true;
-  } else {
-    delete next.plain;
-  }
-
-  saveConfig(next);
-  return next;
-}
-
-function printSavedConfig(vault: string, plain: boolean): void {
-  console.log(`Config saved: ${configPath()}`);
-  console.log(`  vault: ${vault}${plain ? " (plain mode)" : ""}`);
-}
-
-function printObsidianCliStatus(plain: boolean): void {
-  if (plain) return;
-
-  const cli = detectCli();
-  if (cli.installed) {
-    console.log(`  Obsidian CLI: detected (${cli.version ?? "unknown version"})`);
-    console.log(`  Daily notes will be written via CLI when Obsidian is running.`);
-  } else {
-    console.log(`  Obsidian CLI: not detected (using direct file write)`);
-    console.log(`  For better integration, enable CLI in Obsidian ${MIN_CLI_VERSION}+ Settings.`);
-  }
-}
-
-function detectBinary(bin: "agentlog" | "codex"): string {
-  const result = spawnSync("/bin/sh", ["-lc", `command -v ${bin}`], {
-    encoding: "utf-8",
-  });
-  if (result.status !== 0) return "";
-  return (result.stdout ?? "").trim().split("\n").filter(Boolean).pop() ?? "";
-}
 
 type InitTarget = "claude" | "codex" | "all";
 type UninstallTarget = "claude" | "codex" | "all";
